chore(etl): zero-out ruff lint debt (118 errors -> 0)
Mirror di piersoft/cruscotto-italia 67de50a.

Riparazione completa del debito tecnico ruff su tutta la codebase ETL.
Prima di questo commit CI 'lint & test' falliva da quando ruff e' stato
inserito nella pipeline; ora passa pulito.

Sintesi dei 118 errori risolti:
- Auto-fix safe (73): import sorting/unused, f-string vuoti, escape seq.
- Auto-fix unsafe selettivo (13): B007/B905/RUF059.
- Fix manuali (32):
  * F841 (7): codice morto rimosso
  * B023 (6): closure binding con default arg
  * B007 (6): rinomina var loop unused -> _var
  * RUF046 (5): cast int() superfluo
  * RUF002/003 (5): Unicode ambigui (× -> x, ∪ -> U)
  * RUF013/005/UP028 (3): altri pattern moderni

Verifica:
- ruff check . -> 'All checks passed!'
- py_compile su tutti i file modificati: OK
- Smoke test pipeline pun.py: numeri invariati (Roma 3680 PdR)

20 file modificati, -24 righe nette.

+++ etl/sources/istat_turismo.py
@@ -34,12 +34,11 @@ from __future__ import annotations
 
 import argparse
 import json
-import os
 import re
 import sys
 import tempfile
-import urllib.request
 import urllib.error
+import urllib.request
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from pathlib import Path
 
@@ -378,7 +377,7 @@ def build_shard(istat, prov_nuts3, prov_nome, cap_data, fl_data, pop):
 
     def cat_int(data_type, type_acc, ateco):
         v = get_cat(data_type, type_acc, ateco)
-        return int(round(v)) if v is not None else None
+        return round(v) if v is not None else None
 
     # Totali (TYPE=ALL, ATECO=551_553 = somma alberghi+extra)
     totale_strutture = cat_int(DATA_NUM_EST, "ALL", "551_553")
@@ -436,7 +435,7 @@ def build_shard(istat, prov_nuts3, prov_nome, cap_data, fl_data, pop):
     # === FLUSSI: per provincia ===
     def fl(dt, country):
         v = fl_data.get((dt, country))
-        return int(round(v)) if v is not None else None
+        return round(v) if v is not None else None
 
     arrivi_tot = fl("AR", "WORLD")
     arrivi_ita = fl("AR", "IT")
